pyproject.toml
[project]
name = "posprint"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-escpos>=3.1",
    "pywin32>=311",
]
